+++ packages/api-plugin-authorization-simple/src/util/defaultRoles.js
@@ -70,6 +70,7 @@ export const defaultShopManagerRoles = [
   "reaction:legacy:shippingRestrictions/update",
   "reaction:legacy:shops/read",
   "reaction:legacy:shops/update",
+  "reaction:legacy:simpleSchema/introspect",
   "reaction:legacy:sitemaps/update:settings",
   "reaction:legacy:surcharges/create",
   "reaction:legacy:surcharges/delete",
@@ -118,5 +119,6 @@ export const defaultSystemManagerRoles = [
   "reaction:legacy:groups/update",
   "reaction:legacy:shops/create",
   "reaction:legacy:shops/read",
+  "reaction:legacy:simpleSchema/introspect",
   "reaction:legacy:invitations/read"
 ];