+++ packages/api-plugin-fulfillment/src/queries/fulfillmentMethods.js
@@ -1,5 +1,5 @@
 /**
- * @name getFulfillmentMethods
+ * @name fulfillmentMethods
  * @method
  * @memberof Fulfillment/Queries
  * @summary Query the Fulfillment collection for a list of fulfillment Methods
@@ -9,7 +9,7 @@
  * @param {String} input.fulfillmentTypeId - The fulfillmentType id of the fulfillment type
  * @returns {Promise<Object>} Mongo cursor
  */
-export default async function getFulfillmentMethods(context, input) {
+export default async function fulfillmentMethods(context, input) {
   const { collections: { Fulfillment } } = context;
   const { shopId, fulfillmentTypeId } = input;