+++ packages/api-plugin-fulfillment/src/resolvers/Mutation/selectFulfillmentOptionForGroup.js
@@ -12,6 +12,17 @@ import selectFulfillmentOptionForGroupMutation from "../../mutations/selectFulfi
  * @param {String} [args.input.cartToken] - The token for the cart, required if it is an anonymous cart
  * @param {String} args.input.fulfillmentGroupId - The group to select a fulfillment option for
  * @param {String} args.input.fulfillmentMethodId - The fulfillment method ID from the option the shopper selected
+ * @param {Object} context - an object containing the per-request state
+ *//**
+ * The old, deprecated way to call selectFulfillmentOptionForGroup.
+ *
+ * @deprecated (reason: "Avoid using opaqueIds and clientMutationId")
+ * @param {Object} parentResult - unused
+ * @param {Object} args.input - an object of all mutation arguments that were sent by the client
+ * @param {String} args.input.cartId - The ID of the cart to select a fulfillment option for
+ * @param {String} [args.input.cartToken] - The token for the cart, required if it is an anonymous cart
+ * @param {String} args.input.fulfillmentGroupId - The group to select a fulfillment option for
+ * @param {String} args.input.fulfillmentMethodId - The fulfillment method ID from the option the shopper selected
  * @param {String} [args.input.clientMutationId] - An optional string identifying the mutation call
  * @param {Object} context - an object containing the per-request state
  * @returns {Promise<Object>} SelectFulfillmentOptionForGroupPayload