+++ packages/api-plugin-simple-schema/index.js
@@ -1,3 +1,5 @@
 import register from "./src/index.js";
 
+export { default as migrations } from "./migrations/index.js";
+
 export default register;