+++ packages/api-plugin-tags/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @reactioncommerce/api-plugin-tags
 
+## 1.2.1
+
+### Patch Changes
+
+- [#6695](https://github.com/reactioncommerce/reaction/pull/6695) [`1dd0c0468`](https://github.com/reactioncommerce/reaction/commit/1dd0c04687e473489e7225dda8b2b880df1b94b2) Thanks [@zenweasel](https://github.com/zenweasel)! - Allow getSlug to take a second parameter of allowedChars
+
+- Updated dependencies [[`8392ba163`](https://github.com/reactioncommerce/reaction/commit/8392ba163a402ba0528f5658bd2f206cb9433eee), [`1dd0c0468`](https://github.com/reactioncommerce/reaction/commit/1dd0c04687e473489e7225dda8b2b880df1b94b2)]:
+  - @reactioncommerce/api-utils@1.17.1
+
 ## 1.2.0
 
 ### Minor Changes